Ticket #1392: Removed unnecessary catches
MsgProducer::run stores the message back to the disk so it isn't lost, and tries to reconnect later on.
The extra catches just printed the errors, but didn't mark the connection as lost, neither stored back the message on the disk,
so after a necessary restart all those messages would be lost.