FTS-273: Avoid duplicated 'complete' messages
When receiving a signal, SendTransferFinishMessage can be called twice.
The handler calls it, then cancels the transfer and wait a bit.
If the transfer is canceled quickly, then after exiting gfalt_copy_file
another termination message is sent.
Since the signal may be a non-recoverable signal, or the actual cancel
can take long, removing the message from the handler is not really an
option.
This patch introduces a simple state machine to avoid sending twice
the termination message.