FTS-542: When a process is stalled, FAIL READY too
However, we need to update the timestamp on the watchlist for the READY
transfer as well. Otherwise they will expire if the preceding transfer
takes longer than the stall interval.

So we use the pid to update the timestamps, rather than file/job id.

However, since pid may be reused, just to make sure we check the
start time of the process with the last known message. If the process
started after the last message, we infer it is a different process, so we do
not update the watchlist.