FTS-1943: Always create a DstFileReport for all the GFAL_XATTR_STATUS values
The DstFileReport feature relies on the GFAL_XATTR_STATUS reported to determine the file locality. By default, the locality is false for both the onDisk and the onTape flags. If the status is one of the following three, GFAL_XATTR_STATUS_ONLINE, GFAL_XATTR_STATUS_NEARLINE, GFAL_XATTR_STATUS_NEARLINE_ONLINE, FTS sets the file locality as true either for onTape or onDisk accordingly, otherwise it leaves it as false for both.