FTS-1979: Improve the transfer failure error message when token-exchange fails.
Certain token providers (such as IAM) will provide more info in the response body when the token exchange fails. This change tries to interpret the response body as a JSON message and look for the "error" and "error_description" fields. This is tailored for the IAM token provider, which will serve most WLCG communities. If the parsing fails, the generic Davix failed HTTP request message is propagated.

Note: When it comes to the token-exchange request, if it fails because of an invalid access token, IAM will reply with "HTTP 401" status code. This is interpreted by Davix as a failure and the response body will no longer be read. To circumvent this, we replaced the Davix "HttpRequest::executeRequest()" with the "beginRequest()" + "readBlock()" in a loop + "endRequest()" sequence. This will ensure we always read the full response body