FTS-1998: Allow look-up of CloudStorage credentials also for OAuth2 based transfers
CloudStorage credentials look-up is done based on:
  - user identity (certificate DN or token subject), which comes from the transfer object
  - VO attributes, which come from the UserCredential object

OAuth2 submitted transfers don't have an entry in the "t_credential" table, which is what is used during the look-up.
To emulate this, create an empty UserCredential object, which can be used during the search.

Additionally, once credentials are passed to the UrlCopyProcess, ensure the temporary CloudStorage file is deleted