FTS-2028: Assign usernames to XRootd tape tasks, ensuring there will be no clashes between multiple parallel XRootd channels
By default, XRootd will associate a username to a session/channel. When another request arrives in timely manner,
the XRootd library will try to reuse an existing open channel. Given the QoS daemon manages many tasks, including to the same storage,
but with different proxy certificates, we want to avoid this caching behavior