test: PidTools/basic create subprocess to verify getPidStartime()
This test verify that the function getPidStartime() gives a coherent
process start time.
Adding more Boost tests made the test fails, because the time
elapsed in the fts-unit-tests is now longer.

To make this test a bit more coherent, we create a dedicated process to
let the test verify that the starttime is in a "valid" range.

Maybe implementing the getPidStartime() logic in the test can be a good way
to verify that this value doesn't change over time.
Instead of an approach with starttime is in the range of current time -
60s.