test: Fix boost crash, move to dynamic library boost header
The tests of fts links dynamically the Boost library in Cmake. This require to use a specific set of headers
of Boost for the dynamic link use case.

The main unit.cpp was using the Boost header which has to be used in case of including the "library as headers".

This fix a crash happening when running some tests individually:
malloc_consolidate(): invalid chunk size
IOT instruction (core dumped)  fts-unit-tests --run_test=common/PidTools/basic --log_level=all