Modify the TokenExchange service to retrieve tokens from the database on a fixed, configurable limit
This approach prevents a huge workload which would take hours to finish. By retrieving work in small chunks, we allow the system to advance transfers as it works itself through the tokens-to-exchange set. The implementation also keeps track whether the bulk limit was reached in the previous iteration, in order to allow the service to continue work immediately, instead of artificially sleeping