FTS-1990: Implementation for the TokenRefreshPoller service
The service receives token-refresh requests from the TokenRefreshListener service. It will check the set of tokens-to-refresh in the database. Those tokens that have an acceptable lifetime (5min+ remaining) are made available to the TokenRefreshListener service. The others are marked in the database via the "t_token.marked_for_refresh", to later be picked up and refreshed by the TokenRefesh service. In order to avoid overloading the database, the marking of tokens for refresh is done only periodically. The periodic behavior is implemented with the help of an in-memory cache.

The component doing the token-refresh remains to be implemented.

This commit also finalizes the response from the TokenRefreshListener back to the UrlCopyProcess with the refreshed access token