FTS-1990: Implementation for the TokenRefresh service
This service regularly scans the database for access tokens marked for refresh. It will perform the token-refresh flow on the identified tokens. The result is stored back in the database. In case of an error, the "refresh-message" is set with the token-refresh error message. In both update scenarios, the "t_token.marked_for_refresh" flag is reset

The TokenRefreshPoller service will discover these tokens, making them available to the TokenRefreshListener service, which will send the refreshed access tokens back to the calling clients