FTS-1990: Handle the failed token-refresh scenario
The token-refresh workflow is performed and in the case of failure, the error message is stored in the database. This message is later picked-up by the TokenRefreshPoller service, which will make it available to the TokenRefreshListener service:
- Changes were brought to the internal cache of the TokenRefreshPoller service
- When the TokenRefreshListener registers tokens for refreshing, now a timestamp is also associated. This timestamp is used to ensure we send only fresh token-refresh error messages back to the clients. In case the request arrives after the error message was registered in the database, then we treat it as a new request and mark that token ID for refreshing

The structure of the protobuf TokenRefreshResponse was modified to include the error message fields. The UrlCopyProcess also was modified to handle the error scenario