Fix "StoreAString" unit test assert failure
This patch fix the "StoreAString" unit test, here it tries to retrieve
a program parameter "--intpar=10", with the old code it triggers an assert which
verifies that the given program parameter is found.

The way the argc/argv are built for this unit test is error-prone:
The setupParameters() function takes a `const std::string&` parameter,
and assign it to an array of pointer `argv[1]`.

The StoreAString unit test pass a string literal as the parameter.
This results in a temporary std::string being constructed, holding a reference to a copy of the literal string.
The lifetime of this temporary std::string is directly bound with the lifetime of the setupParameters().
The `argv[1]` is then holding a reference to the a temporary std::string which
expires when the program leave the scope of the setupParameters().

$ fts-unit-tests --log_level=all --report_level=detailed --run_test=config/ServerConfigReaderTestSuite/storeAString

```
/fts3/test/unit/config/ServerConfigReader.cpp(287): Entering test case "storeAString"
fts-unit-tests: /work/fts3/src/config/serverconfigreader.cpp:698: void fts3::config::ServerConfigReader::storeAsString(const std::string&) [with T = int; std::string = std::__cxx11::basic_string<char>]: Assertion `isFound' failed.
unknown location(0): fatal error: in "config/ServerConfigReaderTestSuite/storeAString": signal: SIGABRT (application abort requested)
/fts3/test/unit/config/ServerConfigReader.cpp(287): last checkpoint: "storeAString" test entry
/fts3/test/unit/config/ServerConfigReader.cpp(287): Leaving test case "storeAString"; testing time: 320us
```