activemq: Refactor MsgPipe.{h/cpp} into MessageLoader.{h/cpp}
The MsgPipe is loading all monitoring messages from disk/DirQ into a
concurrentQueue. Once it is loaded in memory the message is instantly
deleted from the disk.

This behavior was not ideal, when the fts-msg-bulk lose connection with
the ActiveMQ broker, and the daemon is restarted, all the messages are
lost.

To fix this issue, the MessageLoader will delegate the DirQ message
deletion to the consumer (BrokerPublisher). Which will delete the
message from disk once the ActiveMQ acknowledge the reception.