FTS-2146: Remove the old BringOnline daemon (obsoleted by the QoS daemon in FTS3 v3.10.0)
The following are removed:
- src/bringonline-daemon (main daemon code)
- The "fts-bringonline" systemd unit files
- The "fts_bringonline" manpage

This code folder should've been adapted and renamed into the "fts-qos" daemon.
However, the code was duplicated, creating a suboptimal situation.

Given the FTS4 project on the horizons, the plan to merge the history of the two code folders no longer applies and instead removal is preferable