activemq: use a single configuration file for the fts_activemq daemon
Remove the dependency on the fts_server configuration file.
This requires to move few configuration options into the
fts-activemq.conf:
- User
- Group
- LogLevel
- Alias
- MessageDirectory

The fts_activemq doesn't use Boost to parse command lines options
anymore, and only provide the "--configfile" and "--no-daemon".
Previously, it was showing all the available parameters of the
fts_server, which were no relevant for the fts_msg_bulk daemon.