cmake: backport CMake 4.1 GNUInstallDirs module
This module was patched in CMake upstream and will be available in CMake 4.1. The upstream patch fix an issue where the differents
CMAKE_INSTALL_* directories were not well defined when using CMAKE_INSTALL_PREFIX like "/usr".

FTS supports Almalinux9 which will not see this fix/version of CMake before Almalinux10 or later. Backporting this module
in FTS allows the project to be built in compliance with the FHS (see https://en.wikipedia.org/wiki/Filesystem_Hierarchy_Standard)
especially the "/usr/local" directory.

Once the linux distribution baseline of FTS move to a newer version of Almalinux for example, and this distribution packages CMake 4.1.
Then, we can stop providing the GNUInstallDirs-backport module, since the GNUInstallDirs is distributed with CMake.