cmake: GNUInstallDirs enforce the new directory policy
The CMake GNUInstallDirs module backported in FTS is not directly compatible with
CMake versions older than 4.1. This patch make the module work for
older CMake version and enforce the 'new' policy, which is to bind
correctly the different installation directories.

For example without this patch and the backport of the GNUInstallDirs
module.

$ cmake .. -DCMAKE_INSTALL_PREFIX="/usr"
Set CMAKE_INSTALL_SYSCONFDIR to "/usr/etc" instead of "/etc".