activemq: fix dual broker connection, improve logging
The reverse DNS resolution was giving a ActiveMQ hostname per ipv4 and
ipv6 address. The fts_activemq daemon use only the hostname to connect
to a broker. This commit fix the dual connection per broker which could
happen if the resolved alias contains both ipv4 and ipv6.

Improve the overall logging of the daemon, the daemon was showing state
changes even if the message dispatch doesn't succeed. Which is
missleading compared to the previous fts_msg_bulk behaviors.

Now the deamon logs the dispatch attempt, and an other message once the
Broker acknowledge the reception.