FTS-2210: Uniform handling of OAuth2 tokens, regardless whether dealing with HTTP or Cloud-based transfers.
Effectively, this will allow OAuth2 token-based transfers to work with the new token workflow, which refreshes just-in-time.

There is a disadvantage: the transfer agent manages both the token for the source and the destination, despite not using it for the cloud endpoint. To address this, a larger work item is needed, as we always expect a token-based transfer to have a token for both the source and the destination