FTS-2224: Don't use macaroons to authenticate against the Tape Rest API
Starting with v3.14.0, the macaroon management moved to FTS. Macaroons are retrieved and set in the Gfal2 credentials map in the Gfal credentials mapping. This happens among the first things in the UrlCopyProcess. Once stored in the Gfal2 keystore, before any storage operation, Gfal2 searches and loads any credential associated with the request URL.

When Gfal2 was managing the macaroon retrieval, it avoided getting a macaron for Tape REST API requests. Now that the macaroon is passed-in by FTS, we ended up with the scenario that Gfal2 would issue Tape REST API requests to the destination storage with macaroons.

This commit ensures destination macaroons should be used only for the exact URL, and not when just involving the destination host.